Extract permission form logic to trait
